Add reparse test macro
Brings the .chalk file writer tests in line with the existing chalk
tests which use macros for programs.

This makes the test programs easier to read as they have syntax
highlighting in IDEs, and are no longer surrounded by quotes.
