+++ tests/display/formatting.rs
@@ -1,63 +1,64 @@
-use super::*;
-
 #[test]
 fn test_assoc_type_formatting() {
-    test_formatting(
-        "
-        struct Foo {}
-        trait Bar {
-            type Assoc;
+    reparse_test!(
+        program {
+            struct Foo {}
+            trait Bar {
+                type Assoc;
+            }
+            impl Bar for Foo {
+                type Assoc = ();
+            }
         }
-        impl Bar for Foo {
-            type Assoc = ();
-        }
-        ",
-        r#"struct [a-zA-Z0-9_-]+ \{\s*\}
+        formatting matches
+r#"struct [a-zA-Z0-9_-]+ \{\s*\}
 trait [a-zA-Z0-9_-]+ \{
   type [a-zA-Z0-9_-]+;
 \}
 impl [a-zA-Z0-9_-]+ for [a-zA-Z0-9_-]+ \{
   type [a-zA-Z0-9_-]+ = \(\);
-\}"#,
+\}"#
     );
 }
 
 #[test]
 fn test_struct_field_formatting() {
-    test_formatting(
-        "
-        struct Foo {}
-        struct Bar {
-            field1: Foo
+    reparse_test!(
+        program {
+            struct Foo {}
+            struct Bar {
+                field1: Foo
+            }
+            struct Azg {
+                field1: Foo,
+                field2: Bar
+            }
         }
-        struct Azg {
-            field1: Foo,
-            field2: Bar
-        }
-        ",
-        r#"struct [a-zA-Z0-9_-]+ \{\}
+        formatting matches
+r#"struct [a-zA-Z0-9_-]+ \{\}
 struct [a-zA-Z0-9_-]+ \{
   [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+
 \}
 struct [a-zA-Z0-9_-]+ \{
   [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+,
   [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+
-\}"#,
+\}"#
     );
 }
 
 #[test]
 fn test_where_clause_formatting() {
-    test_formatting(
-        "
+    reparse_test!(
+    program {
         struct Foo where Foo: Baz, Foo: Bar {}
         trait Bar where Foo: Baz, dyn Baz: Bar {}
         trait Baz {}
         impl Bar for Foo where Foo: Baz, (): Baz {}
         impl Baz for Foo {}
         impl Bar for dyn Baz {}
-    ",
-        r#"struct [a-zA-Z0-9_-]+
+    }
+    formatting matches
+r#"struct [a-zA-Z0-9_-]+
 where
   [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+,
   [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+
@@ -74,30 +75,31 @@ where
   \(\): [a-zA-Z0-9_-]+
 \{\}
 impl [a-zA-Z0-9_-]+ for [a-zA-Z0-9_-]+ \{\}
-impl [a-zA-Z0-9_-]+ for dyn [a-zA-Z0-9_-]+ \{\}"#,
+impl [a-zA-Z0-9_-]+ for dyn [a-zA-Z0-9_-]+ \{\}"#
     );
 }
 
 #[test]
 fn test_assoc_ty_where_clause() {
-    test_formatting(
-        "
-        trait Bar {}
-        trait Fuzz {
-            type Assoc
-            where 
-                dyn Bar: Bar,
-                Self: Bar;
+    reparse_test!(
+        program {
+            trait Bar {}
+            trait Fuzz {
+                type Assoc
+                where
+                    dyn Bar: Bar,
+                    Self: Bar;
+            }
         }
-    ",
-        r#"trait [a-zA-Z0-9_-]+ \{\s*\}
+        formatting matches
+r#"trait [a-zA-Z0-9_-]+ \{\s*\}
 trait [a-zA-Z0-9_-]+ \{
   type [a-zA-Z0-9_-]+
   where
     dyn [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+,
     [a-zA-Z0-9_-]+: [a-zA-Z0-9_-]+;
 \}
-"#,
+"#
     );
 }
 
@@ -107,18 +109,18 @@ fn test_name_disambiguation() {
     // structs or traits with the same name in Chalk - but luckily our
     // implementation ignores types for name disambiguation, so we can test it
     // indirectly by using a opaque type and trait of the same name.
-    reparse_into_different_test(
-        "
-        struct Foo {}
-        trait Baz {}
-        impl Baz for Foo {}
-        opaque type Baz: Baz = Foo;
-        ",
-        "
-        struct Foo {}
-        trait Baz {}
-        impl Baz for Foo {}
-        opaque type Baz_1: Baz = Foo;
-        ",
+    reparse_test! (
+        program {
+            struct Foo {}
+            trait Baz {}
+            impl Baz for Foo {}
+            opaque type Baz: Baz = Foo;
+        }
+        produces {
+            struct Foo {}
+            trait Baz {}
+            impl Baz for Foo {}
+            opaque type Baz_1: Baz = Foo;
+        }
     );
 }
