+++ tests/display/lifetimes.rs
@@ -1,8 +1,7 @@
-use super::*;
 #[test]
 fn test_various_forall() {
-    reparse_test(
-        "
+    reparse_test!(
+        program {
             struct Foo<'b> where forall<'a> Foo<'a>: Baz<'a> { }
             trait Baz<'a> {}
             trait Bax<'a> {}
@@ -12,16 +11,16 @@ fn test_various_forall() {
             impl<'a> Baz<'a> for for<'b> fn(Foo<'b>) { }
             impl<'a> Bax<'a> for fn(Foo<'a>) { }
             impl<'a> Bax<'a> for dyn forall<'b> Baz<'b> { }
-            ",
+        }
     );
 }
 #[test]
 fn test_lifetimes_in_structs() {
-    reparse_test(
-        "
+    reparse_test!(
+        program {
             struct Foo<'b> { }
             trait Baz<'a> {}
             impl<'a> Baz<'a> for Foo<'a> { }
-            ",
+        }
     );
 }
